Add possibility to run unit tests
Those tests are based on FreeRTOS-kernel repo. The build system is based
on CMake. After the command:

    mkdir build && cd build && cmake ..  && make

is invoked the Catch2 repo and FreeRTOS-kernel repo are downloaded. The
unit tests are run then on top of Catch2 and at this stage only FreeRTOS
OS is supported. Most probably mbed, or other OSes will be supported.