Fix segmentation fault for os_delayed_job
The problem was that for the timer task, we have passed a pointer to the
shared_ptr to the proxy object (of type os_delayed_job_data). This was
really bad because we passed the address to the member shared_ptr of
os_delayed_job. This means that we passed a pointer to a local member
which could be destroyed after call to detach() function and destructing
the os_delayed_job instance. This means that when the detached callback
was called it captured an address to a nonexisting variable.
To fix that we pass the address of the actual object managed by the
std::shared_ptr, using the ::get() function (std::shared_ptr::get()).
As long as there is a std::shared_ptr which manages the object that
pointer (returned by get()) is valid and we trust that std::shared_ptr
will not invalidate that pointer, then we are good to go with that
bugfix.