Implement an asker
Asker allows to ask a specific predicate with specified intervals, for a
specified amount of time. The asker needs an OS-dependent delayer
function.
Asker asks a predicate for a specified duration. If within that period
of time the predicate returns true asker also returns true, but on
timeout it returns false.