Rework generic active and accomodate native active
Generic active has changed due to that we want to provide references to
implementations of the thread and message pump.
For the native target we simply create a std::thread and a simple
message_pump, but for FreeRTOS we would like to have control over the
constructor arguments we specify to the thread or the message pump.