Adapt thread to have start() function
Some rework had to be done because since now we use thread::start()
function. Because of that we needed to make some changes and move the
thread initialization code to the start() method.
This enables to use std::enable_shared_from_this and get rid of some
comments which mentioned that.