Extract insert() as separate method
insert() is extracted from the send() method. It implements the
part which is not coupled to FreeRTOS.

It is protected so the inheritors could use it internally.

This is done for the inheritors to reuse all the needed common parts
of the code and overwrite the FreeRTOS-related code when using send().