Implement lock free insert()
send_from_isr() was faulty because I missed that the lockguard uses
non-ISR API (FreeRTOS's from-task-API).
Since now, insert_lock_free() inserts the element to the underlying
queue. No lockguard is used.
This is wrong, and it should be fixed. We have to find a way to make it
safer.