Use std::array for underlying circular_buffer
This forces the stored element type to be default constructible, what is
not so important, but may be inconvenient though.

I had to make this change, because with std::vector, there was a bug
when calling insert(). Insert will insert a new element at specific
index, what is wrong. The queue_depot size would then increase with each
element added. This is wrong - it's a memory leak.

One more thing was that the std::vector called operator::new. That could
be called from ISR and this is also wrong.

We use std::array which:

1. Makes the size static, so no way to have an error there with memory
leakage.
2. Is statically allocated, so no calls to the operator::new will be
performed in regards to queue_depot.