Fix max event bits; go full compile-time
With this commit we switch the event_group to be compile-time, so we can
check during compile time, if the events we set, clear, await on, are
within the proper set of the enumerations.

The number of maximum event bits is also calculated. The EventGroupWait
tests failed because we used event bits that cannot be used for 32-bit
TickType_t. With this commit we also static_assert() on too many event
bits supplied.